wip: v1 guided decoding
so far i'm focused on FSM initialization and finding the right place
for that to occur and do it in such a way that it can be async in
another thread.

multi-modal input processing was already implemented and appeared to
follow a similar pattern, so I added the async initiailization logic
around that to start.