Experiment with deferring Guide initialization to a thread
Unlike with AsyncLLMEngine, with MQLLMEngine we cannot suspend the
handling of a request and process other requests while we wait for
Guide initialization.

However, if we use a Future, we should be able to allow the Guide
initialization to be happening in another thread while we are
accepting other available requests.

This assumes that Guide initialization releases the GIL as per
Russell's branch that uses py.allow_threads() in the PyIndex
constructor in outlines-core.