gpu_model_runner: Don't run encoder directly for encoder-decoder
For multimodal models, the encoder is run directly first so that the
result can be reused potentially multiple times when being substituted
into the decoder prompt. This case is not relevant for
encoder-decoder, so it's simpler to just execute the model as usual
and let the encoder run during the first forward() call to the model.
It will only run the first time once everything is working properly.

Signed-off-by: Russell Bryant <rbryant@redhat.com>